Added some required services

+++ src/main/java/com/redit/db/Employee.java
@@ -43,7 +43,7 @@ public class Employee {
         return DaoManager.createDao(con, Employee.class);
     }
 
-    public void addApprovedExpense(int amount) {
+    public void addApprovedExpense(long amount) {
         reimbursement += amount;
     }
